:memo: Restructured file structure by putting modules in /pages
