+++ client/src/pages/shared/context/auth-context.js
